refactor: ProjectDetailLayout Context 분리
